+++ ports/stm32/boards/Passport/modules/tests/test_unit.py
@@ -20,6 +20,10 @@ def test_ext_settings(test):
     assert test('ext_settings.py') == b'OK'
 
 
+def test_psbt_multisig_approval(test):
+    assert test('psbt_multisig_approval.py') == b'OK'
+
+
 def test_ui(test):
     assert test('ui.py') == b'OK'